requirements.txt
playwright==1.48.0
# Updated to a published CrewAI version
# Downgraded CrewAI to avoid python-dotenv version conflict
crewai==1.4.1
groq==0.13.0
python-dotenv==1.0.0
# Updated for CrewAI compatibility (requires >=2.11.9)
# Adjusted to latest available fulfilling CrewAI constraint
pydantic==2.12.4
markdown==3.7
weasyprint==62.3
rich==13.9.4
requests==2.31.0
aiofiles==24.1.0
retrying==1.3.4
# NEW: Serper.dev integration for Pakistani marketplace search
# Additional dependencies for improved error handling
tenacity==8.2.3
# Streamlit UI dependencies
streamlit==1.32.0
# Updated pandas for Python 3.13 compatibility
pandas==2.3.3
plotly==5.18.0
